WIP: Add visiting logic to support OR operation and nested filter conditions
There are two ways of doing this, the first is to support arbitary recursion in the filter operator itself (which will be evaluated everytime when a record is encountered), the other is to 'split' the filter operators into multiple filter operators. We are following the latter. The visiting logic 1. pauses the 'core' visiting logic of the logical/physical plan, 2. constructs these operators, 3. chains the filter operators internally, 4. and finally ties the ends to the 'core' visiting logic. The filter operator is yet to be updated to support OR condition.