Need ordered MatView for planner unittest
For deterministic results, I believe we need the
`FilterSingleORCondition` planner unit test to use an ordered
materialized view. In particular, the test iterates over the keys
in the view and compares the records to records in a vector of
input records, by index. For an unordered MV, the key iteration
order is not defined, so the test can fail nondeterministically.
An ordered MV solves this problem, as the key order is defined.