Replace q29 with q36 (as per benchmark)
This works around the lack of support for nested view definitions by
replacing Q29 (`replying_comments_for_count` + `q29` view over it)
with a merged Q36, as we used in the Lobsters benchmark.

The query currently returns no data, as we don't insert anything into
`read_ribbons`, so the test only verifies that we plan it correctly,
not the output.